requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.107.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
gunicorn==23.0.0

# Production dependencies
redis==5.2.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
opencensus-ext-azure==1.1.13
opencensus-ext-requests==0.8.0
azure-keyvault-secrets==4.9.0
azure-identity==1.19.0
stripe==14.1.0
requests==2.32.3
python-multipart==0.0.20

# Celery for background tasks
celery==5.4.0
kombu==5.4.2

# Azure Speech and Translation Services
azure-cognitiveservices-speech==1.40.0
azure-ai-translation-text==1.0.0

# Testing dependencies
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.25.2
pytest-mock==3.14.0
Pillow==11.0.0
psutil==6.1.1

# Video analysis and computer vision (optional - disabled for Azure deployment)
# mediapipe==0.10.9
# opencv-python-headless==4.8.1.78
numpy==1.24.4

# Analytics and data science
scipy==1.11.4
scikit-learn==1.3.2

# Azure Communication Services (for notifications)
azure-communication-email==1.0.0
azure-communication-sms==1.0.1

# Azure Blob Storage (for video uploads)
azure-storage-blob==12.19.0

# Multi-language support already included above with Speech Services
